laying multiplayer groundwork
i will be building the project with 4-player co-op in mind from the beginning.
it shouldn't be too difficult since i got it working with one Player instance in the Action RPG with some hacky wackiness.
the challenge this time around is going to be allowing any of the 4 players to drop-in & out at any time. in the action RPG, the multiplayer is "toggled" in that it is either on or off. here, each Player should be able to join and exit independently. one thing that will need consideration is how the camera logic will work depending on the number of players.